Rewrite Instance
