+++ Inferno/Graphics/Renderer.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Inferno.Core;
 using Inferno.Graphics.Text;
 
 namespace Inferno.Graphics
