+++ Inferno/Inferno.Runtime.projitems
@@ -9,7 +9,6 @@
     <Import_RootNamespace>Inferno</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Core\Instance.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Game.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Game.Desktop.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Game.UWP.cs" />
@@ -47,6 +46,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Input\Key.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Input\Mouse.Desktop.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Input\MouseState.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Instance.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MathHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Matrix.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Point.cs" />
