+++ Inferno/UI/UserInterface.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using Inferno.Core;
 using Inferno.Graphics;
 using Inferno.UI.Controls;
